requirements.txt
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
httpx==0.28.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.44
supabase==2.24.0
Werkzeug==3.1.3
